edit models

+++ src/main/java/kg/attractor/jobsearch/models/WorkExperienceInfo.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class WorkExperienceInfo {
-    private int resumeId;
+    private Long resumeId;
     private int years;
     private String companyName;
     private String position;
     private String responsibilities;
-    private int id;
+    private Long id;
 }
